requirements.txt
annotated-types==0.7.0
anyio==4.8.0
beautifulsoup4==4.12.3
bs4==0.0.2
cachetools==5.5.2
certifi==2024.12.14
charset-normalizer==3.4.1
colorama==0.4.6
distro==1.9.0
google-api-core==2.25.1
google-api-python-client==2.172.0
google-auth==2.45.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-genai==1.56.0
googleapis-common-protos==1.70.0
gspread==6.2.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
jiter==0.8.2
mysql-connector==2.2.9
mysql-connector-python==9.1.0
oauthlib==3.3.0
openai==1.59.3
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.10.4
pydantic_core==2.27.2
pyparsing==3.2.3
python-dotenv==1.2.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
sniffio==1.3.1
soupsieve==2.6
tenacity==9.1.2
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.2.0
urllib3==2.3.0
websockets==15.0.1
